+++ src/test/java/org/folio/rest/api/StorageTestSuite.java
@@ -120,6 +120,8 @@ public class StorageTestSuite {
     removeTenant(TENANT_ID);
 
     CompletableFuture<String> undeploymentComplete = new CompletableFuture<>();
+ 
+    PostgresClient.getInstance(vertx).stopEmbeddedPostgres();
 
     vertx.close(res -> {
       if(res.succeeded()) {